Move page attribute related logic to page module
